+++ .github/release-notes.md
@@ -1,3 +1,4 @@
 ## Changes:
 
 -   Use whisper.cpp by default on Windows.
+-   New argument: `--model`